[Decomp][Item] Cohesion split: extract enchantment application to ItemEnchantment.cpp
