+++ src/game/Object/Item.cpp
@@ -1229,151 +1229,11 @@ bool Item::IsTargetValidForItemUse(Unit* pUnitTarget)
     return false;
 }
 
-/**
- * @brief Sets enchantment data for a slot.
- *
- * @param slot The enchantment slot.
- * @param id The enchantment id.
- * @param duration The enchantment duration.
- * @param charges The remaining enchantment charges.
- */
-void Item::SetEnchantment(EnchantmentSlot slot, uint32 id, uint32 duration, uint32 charges)
-{
-    // Better lost small time at check in comparison lost time at item save to DB.
-    if ((GetEnchantmentId(slot) == id) && (GetEnchantmentDuration(slot) == duration) && (GetEnchantmentCharges(slot) == charges))
-    {
-        return;
-    }
 
-    SetUInt32Value(ITEM_FIELD_ENCHANTMENT_1_1 + slot * MAX_ENCHANTMENT_OFFSET + ENCHANTMENT_ID_OFFSET, id);
-    SetUInt32Value(ITEM_FIELD_ENCHANTMENT_1_1 + slot * MAX_ENCHANTMENT_OFFSET + ENCHANTMENT_DURATION_OFFSET, duration);
-    SetUInt32Value(ITEM_FIELD_ENCHANTMENT_1_1 + slot * MAX_ENCHANTMENT_OFFSET + ENCHANTMENT_CHARGES_OFFSET, charges);
-    SetState(ITEM_CHANGED);
-}
 
-/**
- * @brief Updates the duration for an enchantment slot.
- *
- * @param slot The enchantment slot.
- * @param duration The new duration.
- */
-void Item::SetEnchantmentDuration(EnchantmentSlot slot, uint32 duration)
-{
-    if (GetEnchantmentDuration(slot) == duration)
-    {
-        return;
-    }
 
-    SetUInt32Value(ITEM_FIELD_ENCHANTMENT_1_1 + slot * MAX_ENCHANTMENT_OFFSET + ENCHANTMENT_DURATION_OFFSET, duration);
-    SetState(ITEM_CHANGED);
-}
 
-/**
- * @brief Updates the charges for an enchantment slot.
- *
- * @param slot The enchantment slot.
- * @param charges The new charge count.
- */
-void Item::SetEnchantmentCharges(EnchantmentSlot slot, uint32 charges)
-{
-    if (GetEnchantmentCharges(slot) == charges)
-    {
-        return;
-    }
 
-    SetUInt32Value(ITEM_FIELD_ENCHANTMENT_1_1 + slot * MAX_ENCHANTMENT_OFFSET + ENCHANTMENT_CHARGES_OFFSET, charges);
-    SetState(ITEM_CHANGED);
-}
-
-/**
- * @brief Clears all enchantment data from a slot.
- *
- * @param slot The enchantment slot.
- */
-void Item::ClearEnchantment(EnchantmentSlot slot)
-{
-    if (!GetEnchantmentId(slot))
-    {
-        return;
-    }
-
-    for (uint8 x = 0; x < 3; ++x)
-    {
-        SetUInt32Value(ITEM_FIELD_ENCHANTMENT_1_1 + slot * MAX_ENCHANTMENT_OFFSET + x, 0);
-    }
-    SetState(ITEM_CHANGED);
-}
-
-bool Item::GemsFitSockets() const
-{
-    bool fits = true;
-    for (uint32 enchant_slot = SOCK_ENCHANTMENT_SLOT; enchant_slot < SOCK_ENCHANTMENT_SLOT + MAX_GEM_SOCKETS; ++enchant_slot)
-    {
-        uint8 SocketColor = GetProto()->Socket[enchant_slot-SOCK_ENCHANTMENT_SLOT].Color;
-
-        uint32 enchant_id = GetEnchantmentId(EnchantmentSlot(enchant_slot));
-        if (!enchant_id)
-        {
-            if (SocketColor) fits &= false;
-            {
-                continue;
-            }
-        }
-
-        SpellItemEnchantmentEntry const* enchantEntry = sSpellItemEnchantmentStore.LookupEntry(enchant_id);
-        if (!enchantEntry)
-        {
-            if (SocketColor) fits &= false;
-            {
-                continue;
-            }
-        }
-
-        uint8 GemColor = 0;
-
-        uint32 gemid = enchantEntry->GemID;
-        if (gemid)
-        {
-            ItemPrototype const* gemProto = sItemStorage.LookupEntry<ItemPrototype>(gemid);
-            if (gemProto)
-            {
-                GemPropertiesEntry const* gemProperty = sGemPropertiesStore.LookupEntry(gemProto->GemProperties);
-                if (gemProperty)
-                {
-                    GemColor = gemProperty->color;
-                }
-            }
-        }
-
-        fits &= (GemColor & SocketColor) ? true : false;
-    }
-    return fits;
-}
-
-uint8 Item::GetGemCountWithID(uint32 GemID) const
-{
-    uint8 count = 0;
-    for (uint32 enchant_slot = SOCK_ENCHANTMENT_SLOT; enchant_slot < SOCK_ENCHANTMENT_SLOT + MAX_GEM_SOCKETS; ++enchant_slot)
-    {
-        uint32 enchant_id = GetEnchantmentId(EnchantmentSlot(enchant_slot));
-        if (!enchant_id)
-        {
-            continue;
-        }
-
-        SpellItemEnchantmentEntry const* enchantEntry = sSpellItemEnchantmentStore.LookupEntry(enchant_id);
-        if (!enchantEntry)
-        {
-            continue;
-        }
-
-        if (GemID == enchantEntry->GemID)
-        {
-            ++count;
-        }
-    }
-    return count;
-}
 
 /**
  * @brief Checks whether the item is restricted to a different map or zone.
